pyproject.toml
[tool.poetry]
name = "learning-project"
version = "0.1.0"
description = "this is a learning project"
authors = ["Ishan Kumar <ishan1701@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
google-cloud-storage = "^2.18.2"
requests = "^2.32.3"
pendulum = "^3.0.0"
kagglehub = "^0.3.5"
kaggle = "^1.6.17"
PyYAML = "^6.0.2"
asn1crypto="^1.5.1"
bleach="^6.2.0"
cachetools="^5.5.0"
certifi="^2024.12.14"
charset-normalizer="^3.4.1"
google-api-core="^2.24.0"
google-auth="^2.37.0"
google-cloud-core="^2.4.1"
google-cloud-bigquery="^3.31.0"
google-crc32c="^1.6.0"
google-resumable-media="^2.7.2"
googleapis-common-protos="^1.66.0"
idna="^3.10"
packaging="^24.2"
pg8000="^1.31.2"
proto-plus="^1.25.0"
protobuf="^5.29.2"
pyasn1="^0.6.1"
pyasn1_modules="^0.4.1"
PyMySQL="^1.1.1"
python-dateutil="^2.9.0.post0"
python-slugify="^8.0.4"
pytz="^2024.2"
rsa="^4.9"
scramp="^1.4.5"
six="^1.17.0"
SQLAlchemy=">=1.4.36,<2.0"
text-unidecode="^1.3"
tqdm="^4.67.1"
typing_extensions="^4.12.2"
tzdata="^2024.2"
urllib3="^2.3.0"
webencodings="^0.5.1"
pylint="3.3.4"
fpdf="1.7.2"
dbt-core="1.9.2"
dbt-postgres="1.9.0"
dbt-duckdb="1.9.2"
dbt-snowflake = "^1.9.2"
duckdb = "^1.2.1"
pytest = "8.3.5"
pytest-mock = "^3.14.0"
mypy = "^1.15.0"
ruff = "^0.11.2"
apache-airflow-providers-google = "^14.1.0"
pyspark = "^3.5.5"
cookiecutter = "^2.6.0"
pydantic = "^2.11.3"
jinja2 = "^3.1.6"
click = "^8.2.0"
pre-commit = "^4.3.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
learning_project = "learning_project.main:cli"

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "T201", # print-statements
]
ignore = []

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = false
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | Taskfile\.yaml
  | requirements\.txt
  | README\.md
  | poetry\.lock
  | pyproject\.toml
)/
'''